Move Gui class to seperate file and rename to GraphicalUi
